third_party/WebKit/LayoutTests/fast/shapes/parsing/parsing-shape-property-aliases.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<script src="parsing-test-utils.js"></script>
<script>
description('Test that CSS Shapes properties with the -webkit prefix alias the unprefixed properties.');

// This test is only necessary until the prefixed CSS properties have been removed

function getAliasCSSValue(alias, property, value)
{
    var element = document.createElement("div");
    element.style.setProperty(alias, value);
    return element.style.getPropertyValue(property);
}

function testPropertyAlias(alias, property, value)
{
    shouldBeEqualToString("getAliasCSSValue('" + alias + "','" + property + "','" + value +"')", value);
}

applyToEachArglist(
    testPropertyAlias,
    [// [alias, property, value]
     ["-webkit-shape-inside", "shape-inside", "rectangle(0px, 0px, 0px, 0px)"],
     ["-webkit-shape-outside", "shape-outside", "rectangle(0px, 0px, 0px, 0px)"],
     ["-webkit-shape-margin", "shape-margin", "10px"],
     ["-webkit-shape-padding", "shape-padding", "10px"],
     ["-webkit-shape-image-threshold", "shape-image-threshold", "0.5"]
    ]
);
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
